Version number is set manually in docs.
The version number for the package is set in several places, setup.cfg,
CHANGELOG.md, src/crispy_forms_gds/__init__.py and docs/conf.py. The
docs version was set by importing the package and the others were set
manually - using a Makefile target. Having a dependency in the docs
is a little wasteful given the other versions are set via make so the
conf.py was updated to remove the package import and instead it will
be managed by the local makefile as well.