Added integration for sphinx.
Setup the inegration between sphinx and setuptools by adding a section
to setup.cgf. Just the source and build dirs were given as the project
name and version numbers will be picked up from conf.py. The build
directory was chosen to be the same as the one used when building using
the makefile in the docs directory so the generated HTML is always in
the same place (the default was to generate it in the build directory
in the root - where the packages are built).

Added a docs target to the makefile to trigger building the HTML docs
using setup.py.

Changed the release in docs/conf.py to be set separately from version.
Ideally version is just the major and minor release numbers whereas
release is the full version number, e.g. 1.2.3b1, etc.