Fixed errors in makefile when specifying commands.
Fixed the paths to pip so pip3 is used instead of the system default,
pip (usually python 2.7 on Ubunut systems). That fixes an error when
trying to in stall the requirements - Django 2.2 is not available for
python 2.7.

Also updated the serve target to run any migrations. It's wasteful
but it's the easiest way of keeping things up to date.

Now you can just clone the repository and run `make serve` to build
everything and run the demo site.