The required attribute on fields is now disabled on the form.
When the code was initially created the required attribute - added
whenever a field was required - was disabled when a field was
rendered. This was done to turn off the popover that the browser
adds if a field is left blank when a form is submitted. It turns
out that this can be controlled on a Django form so the code was
removed for now until it's clear how this behaviour can be best
handled - a custom form class, something else or simply left as-is
so there's no magic going on that might cause confusion.