Removed support for setting the field width.
The width of a text field was set through the 'width' asttribute. However
the initial implementation added support almost globally rather than just
on text fields. Also a better implementation might be to add a custom
Widget or Field class so the scope can be defined precisely. Until then
support was removed - it's simple enough to add the class directly until
the bette rsolution is added.