Removed tests where we're just testing adding attributes.
Removed the tests that checked whether the spellcheck attribute could
be added to a field. They were added as we wanted to have a set of tests
that covered all the material in the GOV.UK Design System component
guide but the reality is that it is just setting an attribute on the
field which the basic functionality in a Widget or a Field.