Restored default mapping of widget class names to css classes.
In the code for the crispy_field templatetag there is a table that maps
the Widget class name to a CSS class. The entry for 'textinput' was
modified to map to the 'govuk-input' css class. This translation can also
be done by a table added to the settings so the default table was restored
until we make a final decision on how to handle the mapping. There are
three possibilities:

1. Keep the mapping in the settings. This requires the user to add this
   to their settings file.

2. Re-add the change to the default map. Since this code is copied from
   django-crispy-forms and the templatetag is registered with another
   name, `crispy_gds_field' it won't interfere with any code that uses
   another template pack.

3. Add the 'govuk-input' class as a default class to custom Widgets or
   Fields.

Which one is the best option right now is not clear. Most likely is re-
adding the mapping to the default table since that involves the least
amount of work by the user.