Fixed bug when added widget css class to existing classes.
When a field is rendered there is a converters table where the class name
of the field's widget is used to look up the CSS class(es) that will be
added to the list of classes given in the widget's attrs dict. If the
name of the conversion class is a prefix of one of the attr classes the
coversion class will not be added.

For example, if the CSS class obtained from the conversion table is
'textinput' and I add the class 'textinput-wide' to the Widget attrs
dict then the 'textinput' class is not added.

The solution is to use split() to break the css classes from the
conversion table and widget attributes into two lists then add the
classes from one list to the other if it is not already present.