Added support for error summaries.
For each form, an Error summary component must be shown in addition to the
errors messages for each field. The summary is displayed at the top of the
page BEFORE the first H1 element. That means that the Error summary is
displayed OUTSIDE of the form. Previously it was assumed the error summary
was displayed at the top of the form.

To fix this, the FormHelper attribute, form_show_error_summary, was renamed
to form_show_non_field_errors and given a default value of False. The error
template was updated to show the non-field errors when this attribute is
True. The assumption is that pages will be viewed by the public so an Error
summary will be used and so we want to avoid displaying non-field errors
twice. Setting this attribute to True when an Error summary is not used on
a page still allows the "traditional" display of non-field errors at the
top of the form.