CRISPY_CLASS_CONVERTERS setting is no longer used for initial configuration.
The field template tag was updated to map the widget class names to the set
of CSS classes needed to render the respective wiget as a Design System
component. As a result the CRISPY_CLASS_CONVERTERS setting is no longer
needed as part of the basic configuration of the template pack and can be
used as it was originally intended as a way of overriding the CSS classes
used.