Added class method to Field to generate specific field types.
Add a simple set of class methods to Field to make it easy to generate
form where the fields have specific attributes, e.g. spellcheck="false".
That not only reduces the amount of typing but sures that the correct
accessibility attributes are added and not forgotten.

There is a general issue that adding some patterns trigger validation
by the browser which will require some feedback to be given to the
user so they know what the problem is. That subject needs to be looked
at in more detail before an extensive set of methods can be added.