Renamed the 'fields' section of the demo site to 'components'
Now that the HTML class made it easy to render the snippets of non-field
components such as Inset text it makes more sense to model the demo site
more closely to the GOV.UK Design System site. The Fields section was
renamed Components to make it easy to make the connection between the
Design System and how to implement things in Django. The app and paths
were also changed so the names of everything are consistent.