Fixed failing tests when running tox with Django 2.2.
The tests were failing when running with Django 2.2 as parse_html in
Django 3.0 was removing extra spaces from the templates in the class
attribute but 2.2 was not. Updated the templates and now the tests
pass in all environments.

Removed 'make install-packages' from the test command when running tox
since it was an old target and is not needed in any case. Now we just
rely on tox building the sdist and install that into the environment.

Changed the Makefile so 'make tests' now runs the tests for a single
enviroment: python 3.8, Django 3.0, but now using tox rather than
pytest. Previopusly we were installing the package into the
development venv and running pytest. It sort of worked but caused all
kinds of headaches trying to test local changes.