Moved test for fields into the tests for layouts.
The tests for fields were checking how the form elements were rendered and
existing layout tests were doing the same thing for the status components
such as tabs. It makes more sense to have them altogether in the layout
tests and that frees up tests/fields to deal with the custom field added
for the Date input component.