Examples with fields now display submitted values.
Updated the various examples with input fields so the values entered
are displayed. The user can return to the form by clicking on a
"continue" button.

This shows the user that the form submission and validation works and
it also functions as a set of manual tests to verify the fields are
rendered correctly and avoids having to add something like selenium
to verify this.