Replaced the Classes section with a Reference section.
The classes section was organized in the same way as the code. That sort
of worked but was not quite right. Changing the name to "Reference" made
the purpose much clearer. For example the settings documentation was
pulled in from the "Get started" section. Now the section is purely
reference.

Readers come here to see the details of the code. The contents
are quite different in nature from the Components section so this also
eliminates the overlap between the two.