Added Header, Footer, Skip link and Phase banner components.
Added a base template, which is based on the production-ready template
available from the Design System site. The template contains the HTML
to implement the Header, Footer, Skip link and Phase banner components.
The components and parts of the components are wrapped in {% block %}
tags to make it easy to use the template in various project with the
respective blocks customised as required.