Added support for adding css classes to HTML.table.
Also added a separate page for Table to the Components section of
the docs. Previously the layout of a table was indirectly covered
in the docs for Tabs. Now that css support has been added and the
other page level elements recently added were also deocumented it
makes more sense to have a separate page.