Disabled isort in testing as there are conflicts with black.
Black insists that pytest imports are a third-party:

    import pytest
    from crispy_forms_gds.lauout import HTML

But iSort insists that 'import' is separated from 'from'

    import pytest

    from crispy_forms_gds.lauout import HTML

The "black" form is set when the file is saved and that triggers a
test failure. Since we are also running iSort on file saves we are
going to leave that in place and remove the checking in the tests.