Added a Choice class to make setting hints and dividers easier.
Previously the classmethods Field.checkboxes() and Field.radios()
had a hints attribute that took a dictionary with the field values
as keys and the strings to be displayed as hints as values. A
Choice class was created so the hint could be defined when the
form was created. That made it a lot easier to create the form with
the code taking a more declarative style which made it easy to see
what was going on. The hint attributes were removed from the
classmethods as a result.