Simplified webpack configuration.
The webpack config file and associated package list were copied over
from another project. Consequently the configuration was more complex
that was needed for the demo site. The configuration was stripped
down to the bare essentials so it will be easier to adapt for any
project that wants to use the template pack.