FormHelper now correctly translates Sizes into css classes
The default label and legend sizes are documented as being set using
values from the Size class, Size.MEDIUM ("m"). When the form is
rendered the sizes are converted to the respective CSS classes, e.g.
govuk-label--m or govuk-fieldset__legend--m. The conversion was not
being peformed so the raw Size value was output in the HTML. This
was not caught as the tests set the CSS class when it should have
set the Size value instead.