Fix bug when compressing empty values in DateInputField
Currently when compressing empty values in DateInputField
and empty list is passed to `.compress`, `.compress` expects
a list of at least 3 values. So this raises an error.

This commit passes a list of 3 `None` values instead which
`.compress` handles appropriately.

It also updates the CI to use the correct requirements and
python env.