test: update list of supported environments for tox
All python versions from 3.6 - 3.11 are supported. There were issues
with pip for 3.12 when running tests so that was left out for now.

Django 3.2 LTS is the lowest version of Django now supported. Django
4.2 LTS is the highest, for now. Django 5.0 will be added once the
python 3.12 pip issues are sorted.

All the django-crispy-forms releases for 1.x, starting from 1.09 are
supported - some porting or a new release is required for 2.x